Back-end/requirements.txt
absl-py==2.1.0
alembic==1.14.0
aniso8601==9.0.1
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cryptography==43.0.1
cycler==0.12.1
deepface==0.0.93
dlib==19.24.6
face-recognition==1.3.0
face-recognition-models==0.3.0
facenet-pytorch==2.6.0
filelock==3.16.1
fire==0.7.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-MySQLdb==2.0.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.12.23
fonttools==4.55.3
fsspec==2024.12.0
gast==0.4.0
gdown==5.2.0
google-auth==2.37.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.1.0
grpcio==1.68.1
gunicorn==23.0.0
h5py==3.12.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==2.11.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
lz4==4.3.3
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
mtcnn==1.0.0
mysqlclient==2.2.4
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pillow==10.2.0
protobuf==3.19.6
psutil==6.1.1
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.0.2
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==13.9.4
rpds-py==0.20.0
rsa==4.9
scipy==1.14.1
seaborn==0.13.2
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.35
sympy==1.13.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-cpu==2.18.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
tf-estimator-nightly==2.8.0.dev2021122109
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.53
ultralytics-thop==2.0.13
urllib3==2.3.0
Werkzeug==3.0.4
wrapt==1.17.0
zipp==3.20.2